fix(cloudrun): add region tags
